+++ packages/dso-toolkit/src/index.ts
@@ -51,6 +51,8 @@ export * from "./components/contact-information";
 
 export * from "./components/context";
 
+export * from "./components/cursor-tooltip";
+
 export * from "./components/date-picker";
 
 export * from "./components/definition-list";